+++ project-base/src/Shopsys/ShopBundle/Model/Category/Category.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Category;
+
+use Doctrine\ORM\Mapping as ORM;
+use Gedmo\Mapping\Annotation as Gedmo;
+use Shopsys\FrameworkBundle\Model\Category\Category as BaseCategory;
+use Shopsys\FrameworkBundle\Model\Category\CategoryData as BaseCategoryData;
+
+/**
+ * @Gedmo\Tree(type="nested")
+ * @ORM\Table(name="categories")
+ * @ORM\Entity
+ */
+class Category extends BaseCategory
+{
+    /**
+     * @param \Shopsys\ShopBundle\Model\Category\CategoryData $categoryData
+     */
+    public function __construct(BaseCategoryData $categoryData)
+    {
+        parent::__construct($categoryData);
+    }
+
+    /**
+     * @param \Shopsys\ShopBundle\Model\Category\CategoryData $categoryData
+     */
+    public function edit(BaseCategoryData $categoryData)
+    {
+        parent::edit($categoryData);
+    }
+}