+++ project-base/src/Shopsys/ShopBundle/Model/Category/CategoryData.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Category;
+
+use Shopsys\FrameworkBundle\Model\Category\CategoryData as BaseCategoryData;
+
+class CategoryData extends BaseCategoryData
+{
+}